Messy code to look up structs
Turns out type info is sufficient. We don't really need to resolve types
to some canonical unique pointer; we have the typename, and if we really
needed to we even have package info. Needs a bunch of cleanup to thread
less information through tracing, and to do everything in a single pass
again.