Mostly sort out statements/expressions
Almost everything is an expression in Coq even if it was a statement in
Go. Additionally, there is a new concept of a binding, which is mostly
an expression but has a name. The last binding in a sequence needs to be
anonymous or it is ill-formed (it names a value that can't be used).