Integrate if expressions
Needed to re-organize blocks a bit (essentially blocks - namely
sequences of binders - need to be expressions in order to nest if/for
expressions correctly). Ultimately can't really enforce the right
recursive structure, but that would boil down to distinguishing between
where a sequence of expressions is allowed and where a single expression
is expected, and this isn't necessary. Worst case we check the entire
function body for structure, flattening if desired.