Support for tracking spawned threads
Unfortunately this makes it hard to avoid capturing loop variables. Go
statements can easily be function literals that take a loop variable and
immediately pass it, capturing the loop variable. The same doesn't work
for passing a first-class function literal to an ordinary function.

The one approach I can see working is to require re-declaring loop
variables with i := i in the loop; ideally this would only be required
if the loop has a spawn expression.