Add def info for all variables to handle shadowing
This now correctly handles shadowing a pointer-wrapped variable but no
longer using a pointer, in which case in the inner scope the variable
should not be dereferenced.

We use this to correctly model this pattern in Go:

```
for i := 0; i < 10; i++ {
  i := i
  go func() {
    fmt.Println(i)
  }
}
```

The outer i is shadowed by the inner i, to ensure that the closure
captures the current loop iteration and not something it reads when it
runs.

Previously, this example was a special case where goose demanded the `i
:= i` to avoid the surprising, impure behavior of the loop variable.

Now, we correctly model the buggy code without the i := i and correctly
model the i := i as dereferencing the loop variable immediately and
passing it as a local to the spawned thread. In the model, this line of
code looks like `let: "i" := !"i" in`, clarifying what's happening.

Along the way, I discovered the old code was incorrect and the model did
not capture this bug. Shadowing the loop variable means that the i = i +
1 assignment doesn't do the right thing; you actually have to use a post
expression to write this code correctly.