Add primitive array support
Still unsound due to not copying arrays when passed, and heavily
under-tested (with no unit tests).

The way arrays are integrated into the type system doesn't require a
length in the type, so it's instead provided to the array functions.
This unfortunately means that the types in go_lang don't include
lengths, even though they are known. We may want to add them to the type
as an unused parameter (although this makes typechecking annoying since
a length would be required but not present in the code).