Fix how external struct types are handled
Fixes bug where an invalid Gallina names along the lines of marshal.Enc
was used for the type. The correct output is struct.t marshal.Enc.S,
treating it as a struct with a qualified name.

Moved a bunch of support from the syntactic (*Ctx).coqType function to
the type-based (*Ctx).coqTypeOfType, so type support should be a little
more consolidated. We could probably take another look at removing
coqType and always operating on the types.Type, with a little more
support for detecting Named (defined) types wrapping primitives.