Support loading embedded fields from non-addressable struct value;
as a subtlety, this switches over to "addressable" loading once it
encounters a single pointer walking up implicit embedded fields. This is
to avoid 't load an entire struct from the heap where it is not needed
(e.g. it should *not* be a data race to mess with separate fields
concurrently).