Don't distinguish untyped and typed bools; handle binary ops better; try
to emit a complete expresssion for constants when possible. If a
constant expression uses an unsupported binary operator, emit the final
literal value instead.

The distinction between typed and untyped int/string is that the typed
versions are `val`s and the untyped ones are necessarily constants and
are thus Gallina `Z`/`string`s. For booleans, however, this is not true.
The result of an equality comparison is an untyped boolean, which means
that Goose cannot map untyped bools to Gallina bools. A bit of hacky
code to make untyped bools and bools translate to the same thing in
Goose.