Emit function to "define" global variables (i.e. allocate them) and
another to initialize them, except for `init()` functions and except for
initializing imported packages. Initializing imported packages requires
that `initialize'` only run once, which could use the globals map but
requires changes to the semantics of GlobalGet to permit reading an key
that hasn't been initialized as a means of checking if a package has
already been initialized.