Add support for struct generics to translator. (#61)
* Add support for struct generics to translator.
I did this by translating struct descriptors for
structs with type parameters to be a function that
takes the go types and returns a descriptor with
the types resolved. This required replacing where
we store the name of the struct with a call to
this function. This should not change how non-generic
structs are translated.

* Fix previous commit comments from PR 61 and other
discovered issues:
1. Made it so we separately handle going from ast
struct to coq type and coq type to string and do
it properly
2. Handled receiver types that are struct generics
3. Handled passing type params to functions for
non-atomic structs
4. Got to the point where both the unit test here
and the channel model compile in perennial
5. Added unit tests for the issues that were found
when I translated the channel model