proofgen: generate new definitions needed for initialization support (#65)
This PR adds generation of PkgInfo, PkgIsDefined, and PkgIsInitialized instances. It also shifts all the WpFuncCall and related instances to be based on is_pkg_defined to be systematic.

At the same time, we change `pkg_name'` to be the name of the package itself. There is some subtlety here: if a package has an identifier of the same name (e.g., package `raft` has a type `raft`, or package `main` has a function `main`), then it must be referenced with a package qualifier.

---------

Co-authored-by: Upamanyu Sharma <upamanyu@mit.edu>