Add proofgen support for generic structs (#92)
Generic types were already being correctly translated as desired by goose; this PR adds proofgen support (which is more or less necessary for structs to be usable). It doesn't add the rest of the generics support we need, for generic functions and methods on generic structs.

Generic structs are translated somewhat differently. I will use the term "generic struct" for two cases: the obvious case of a Named type with type parameters, and a struct that contains any instantiated generic type; they both have to be handled specially.

A non-generic struct Foo is translated to `Foo : go_type` in goose and `Foo.t : Type` (a Gallina record) in proofgen.

A generic struct Foo produces `Foo : val` in goose and now two things in proofgen: `Foo.ty : go_type -> go_type` (or however many type arguments there are, which could be zero) to model the type constructor itself and ``Foo.t `{IntoValTyped T' T} : Type`` to model values of the struct. Notice that the model type is parameterized by a combination of a go_type T and a Gallina Type T' for each type parameter in the original struct. A PureWp relates `Foo` to `Foo.ty`.

Surprisingly the automation for non-generic types more or less works unchanged for this new setup.