Add a suite of basic examples for testing channel specs. (#90)
Add a suite of basic examples for testing the specifications of the channel model.
This includes both minimal toy examples:

Using send for synchronization
Using receive for synchronization
Just send a message and check its value
A single stage 3 value ordered pipeline
A single stage 3 value unordered parallel pipeline
Broadcast notification via close
and a couple of subtle tests:

Coordinated close with multiple senders
Select statement where a receive case sets to nil after close. Make sure we don't select either case twice. The details are described in the tests comments but generally I want to make sure we handle both that pattern and can specify a limited number of sends/receives on a channel that isn't closed.
I added a check with panic at the end so that I won't be able to prove a triple that doesn't meet the desired property. Another option would have been to just have specs with a postcondition that captures the desired property but this way i can use these both as a test and as a triple with {True} C {True} that can't be proven unless the tested properties are correct.

The tests that don't have translation support yet are disabled under goose with a build tag

I also added a test that just calls all these functions to make sure they don't panic in real go code so I don't try to prove something that isn't correct unintentionally