Add proof-setup utility to start proofs (#99)
Generates the right imports (based on dependencies and FFI), the right Context line, and lemmas for all WPs with the right general shape for function and method parameters.

Currently this tool outputs a skeleton file to stdout that is intended for copy-pasting. Eventually it might read an existing proof and report warnings or make automatic fixes.