Better support generics with known underlying type
If a generic function has an input `T ~[]E` and then an input `v T`,
slice operations should be available on `v`, and if the element type is
needed, it is known to be `E`. This commit improves that support,
including some (but probably not all) cases of the same issue for maps
and channels.