Allow translation to use direct calls
Instead of going through the global function/method name -> code
mapping, directly call the function. This is only intended for
bootstrapping (which is currently only the channel model), and only works if all of the code a package calls is also
direct calls (either trusted models or also translated with direct_calls
= true). It creates the limitation that the code must be in topological
order, and means recursion no longer works.